Documentation and halfway negamaxing

+++ TTT_c/game.h
@@ -3,6 +3,6 @@
 
 //int randomMoveMaker(boardState *board);
 void resetBoard(boardState *board);
-void showBoard(int square[9]);
+void showBoard(uint8_t square[9]);
 int gameLoop(boardState *board);
 int init();
